+++ sdk/core/azure-core/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Features Added
 
-- Added generic `ServerSentEvent<T>` and `CloseableIterableStream<T>` APIs for service-specific server-sent event
+- Added generic `ServerSentEvent<T>` and `ServerSentEventListener<T>` APIs for service-specific server-sent event
   implementations.
 
 ### Breaking Changes